refactored code to share styles across components
