+++ client/src/index.jsx
@@ -8,41 +8,8 @@ import QuestionList from './components/qa/QuestionList.jsx';
 import { Overview } from './components/Overview/Overview.jsx'
 import {ProductContext} from './contexts.js'
 import {sampleProduct} from './components/Overview/sampleData.js';
+import {Container, Section, Heading1, Heading2, theme} from './components/Styles/LayoutStyles.jsx';
 
-const theme = {
-  colors: {
-    primary: 'blue',
-    secondary: 'steelblue',
-    background: '	#f0f8ff',
-    text: '#5A5A5A',
-  },
-  fonts: {
-    main: '"Helvetica Neue", sans-serif',
-  },
-};
-
-const Container = styled.div`
-  display: grid;
-  width: 100%;
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 20px;
-  background-color: ${(props) => props.theme.colors.background};
-  color: ${(props) => props.theme.colors.text};
-  font-family: ${(props) => props.theme.fonts.main};
-`;
-
-const Section = styled.section`
-  margin-bottom: 40px;
-`;
-
-const Heading1 = styled.h1`
-  color: ${(props) => props.theme.colors.primary};
-`;
-
-const Heading2 = styled.h2`
-  color: ${(props) => props.theme.colors.secondary};
-`;
 
 const App = () => {
   const [product, setProduct] = React.useState(sampleProduct);
@@ -62,6 +29,7 @@ const App = () => {
           <Overview/>
         </Section>
         <Section>
+          {/* secondary color for heading */}
           <Heading2>Questions & Answers</Heading2>
           <QuestionList />
         </Section>
